### 2022-01-28 #4108 @whedon
[PRE REVIEW]: CrysalGrowthTracker: A Python package for calculating the face advancment rates of crystals from videos taken using synchrotron radiation
**Submitting author:** @jonathanHuwP (<a href="http://orcid.org/0000-0001-5283-8065">Jonathan Pickering</a>)
**Repository:** https://github.com/jonathanHuwP/CrystalGrowthTracker
**Branch with paper.md** (empty if default branch): joss-submission
**Version:** v0.9
