### 2022-01-30 #4108 @arfon
@jonathanHuwP – thanks for your submission to JOSS. The author affiliation strings need to be comma-separated (while being quoted, e.g.:

```
authors:
  - name: Joanna Leng^[co-first author]
    orcid: 0000-0001-9790-162X
    affiliation: "1,3"
  - name: Jonathan H. Pickering^[co-first author]
    affiliation: "2,3"
    orcid: 0000-0001-5283-8065
  - name: Sven L. M. Schroeder
    affiliation: "4,5,6,7"
  - name: Gunjen Das
    affiliation: "4,5,6,7"
```

Could you fix up your paper and ask @whedon to recompile with `@whedon generate pdf from branch joss-submission`
