### 2022-04-04 #4108 @lucydot
@mikapfl @marcocamma @kedokudo @maurov are any of you available to review this submission? I can see you all have expertise in X-ray spectroscopy.

@jonathanHuwP can you confirm that yourself of one of the other authors will be available to participate in the review process over the next month-6 week period?
