### 2022-04-04 #4108 @mikapfl
This software is not dealing with X-ray spectroscopy as far as I can see. At least the README talks about X-ray shadowgraphs, i.e. real-space images. My expertise is in X-ray scattering, i.e. reciprocal space images, with side knowledge about spectroscopy, i.e. frequency space images. I don't know enough about real-space images to review a software dealing with it meaningfully.
