### 2022-04-04 #4108 @maurov
@lucydot thanks for the invitation to review this software, but as pointed out by @mikapfl, it is not dealing with X-ray spectroscopy, so I am not able to review it neither.
