### 2022-04-05 #4108 @lucydot
@CFGrote @stuartcampbell are either of you available to review this submission for JOSS? The one-line project description reads "This project extracts data on the growth rates of individual faces from x-ray video shadowgraphs of growing crystals." - it may be in or around your area of expertise?
