### 2022-04-21 #4333 @editorialbot
[REVIEW]: CrysalGrowthTracker: A Python package for calculating the face advancment rates of crystals from videos taken using synchrotron radiation
**Submitting author:** @jonathanHuwP (<a href="http://orcid.org/0000-0001-5283-8065">Jonathan Pickering</a>)
**Repository:** https://github.com/jonathanHuwP/CrystalGrowthTracker
**Branch with paper.md** (empty if default branch): joss-submission
**Version:** v1.5.0
