### 2022-05-05 #4333 @KedoKudo
@lucydot I did my first round of review and here are my questions for the authors:

- Is it possible to rename the default branch from `master` to `main`?

- After going through the git history and the manuscript, the contribution of Sven Schroeder and Gunjen Das is still unclear to me.  Perhaps the author could provide a little bit more information on their contribution to the project.

- Where can I find some example data to test the software?  The unit test runs fine, but it is difficult for me to evaluate its functionality and performance without access to some testing data.

- It would be great if the author can provide some example usage to show users how to use the software.  I would recommend providing a few screenshots in the tutorial document if applicable.

- I couldn't find any community guidelines within the repository, perhaps the authors forget to push it to remote?

- Is there any other software that provides similar functionality within the field? If so, could the author provide a brief summary on how CrystalCrowthTracker differs from the rest?
  - Is the growth rate extracted from the series of 2D projections converted to the corresponding 3D one?  The speed calculated from 2D projections is often different from the 3D counterparts.

- Here are  two typos (probably) I found in the manuscript
  - At L33, is the question mark after `riverbank` intentional?  Also, how does riverbank related to PyQt5?
  - At L60, is `comer separated value` supposed to be `comma separated value`?
