### 2022-05-10 #4333 @CFGrote
I started the review over at https://github.com/jonathanHuwP/CrystalGrowthTracker/issues. I'll mention all submitted issues here for reference
