### 2022-05-10 jonathanHuwP/CrystalGrowthTracker#20 @CFGrote
JOSS REVIEW: README.md
- [x] **Developed with**: Regargind linux and/or mac systems. I have built the package on Debian 11 Gnu/Linux so far tests returned "Ok", so should not be a problem. I'm happy to share my experience so you could change this to "was built and tested on Windows10, Mac, and GNU/Linux".

- [x] **Quick Start...**: I suggest to use proper code markup such as
```console
git clone
```
looks nicer in the online version than inline codes markup.

- [x] **Changing the environment**: I don't see the purpose of this paragraph. You explain how to remove the env and then reinstall as in the previous section "Quick start". If you mean to explain updating an existing env, you should probably mention
```console
conda env update --name cgt --file environment.yml
```

- [x] **~~Runing~~Running the software**:
    - [ ] Fix typo in title

- [x]  **UnitTesting**: Include a code snippet such as
```console
python tests/testio.py
```
Would also be helpful to have one master test module to run all tests at once. Consult the unittest documentation for how to setup a test suite from test classes.

I'm very pleased with all the details given regarding how to setup IDEs, how to work around QT issues etc. Good job!
