### 2022-05-10 jonathanHuwP/CrystalGrowthTracker#21 @CFGrote
JOSS REVIEW: Building and Testing on GNU/Linux
As mentioned in #20, I'm here sharing my steps taken to build, install, and run unittests on a Debian 11 GNU/Linux system under anacondo.

## Environment
* Followed instructions in README.md to install and activate the environment.

## Installation
* Added a "setup.py" to the repo

works:
```
# setup.py
from setuptools import setup

setup(
    name='CrystalGrowthTracker',
    version='1.0.0',
    packages=['cgt',
             ],
    license='Apache2',
)
```
Then:
```console
pip install -e .
```
installs `cgt` into the conda env.

## Testing
I ran
```console
$> for t in tests/*.py; do python $t 2>&1 | tee $t.log; done
```
which produced the following output:

```
Test video written to /home/grotec/Repositories/JOSSReviews/CrystalGrowthTracker/test_video.avi
test_write_read (__main__.TestIO)
test output ... ok

Ran 1 test in 0.007s

OK
..
Ran 2 tests in 0.003s

OK
....
Ran 4 tests in 0.003s

OK
..
Ran 2 tests in 0.002s

OK
test_goto_end (__main__.TestVideoControls)
test the jump to last frame ... ok
test_goto_frame_box (__main__.TestVideoControls)
test goto spinbox works ... ok
test_goto_frame_button (__main__.TestVideoControls)
test goto button works ... ok
test_goto_start (__main__.TestVideoControls)
test the jump to first frame ... ok
test_initial_state (__main__.TestVideoControls)
test initialized ok ... ok
test_one_frame_backward (__main__.TestVideoControls)
test the controller emitts the the step back signal ... ok
test_one_frame_forward (__main__.TestVideoControls)
test the controller emitts the step forward signal ... ok
test_play_backwards (__main__.TestVideoControls)
test backwards button works ... ok
test_play_forwards (__main__.TestVideoControls)
test forwards button works ... ok
test_play_pause (__main__.TestVideoControls)
test pause button works ... ok
test_slider (__main__.TestVideoControls)
test slider works ... ok
test_zoom_box (__main__.TestVideoControls)
test zoom spinbox works ... ok

Ran 12 tests in 2.018s

OK
```
The last test ("videosourceffmpegtest.py") seemed to take a long time and did not
produce any log. Pressing the [Enter] key results in:

```
Traceback (most recent call last):
  File "tests/videosourceffmpegtest.py", line 105, in <module>
    test_video_source()
  File "tests/videosourceffmpegtest.py", line 45, in test_video_source
    test_file = make_test(test_file)
  File "/home/grotec/Repositories/JOSSReviews/CrystalGrowthTracker/tests/maketestvideo.py", line 111, in make_test
    return save_video(tmp_path, frames_per_second, dir_path)
  File "/home/grotec/Repositories/JOSSReviews/CrystalGrowthTracker/tests/maketestvideo.py", line 87, in save_video
    command.run(capture_stderr=silent)
  File "/opt/anaconda3/envs/cgt/lib/python3.8/site-packages/ffmpeg/_run.py", line 325, in run
    raise Error('ffmpeg', out, err)
ffmpeg._run.Error: ffmpeg error (see stderr output for detail)
```

Debugging, I find that execution never returns from `save_video()` in "maketestvideo.py". Somehow I must have managed to create the output file "test_video.avi" in my CWD and ffmpeg does not like this. I'd suggest adding a `if out_file.is_file()` in `save_video()` to catch this and instruct the user to delete the already existing video. Other solutions may work as well. And make sure in `test_video_source()` that the video does not exist.

Removing "test_video.avi" and running "videosourceffmpegtest.py" again returned "Ok".

## TODO:
- [x] Add setup.py to make installable via pip
- [x] Cleanup/catch test artifacts
