### 2022-05-10 jonathanHuwP/CrystalGrowthTracker#22 @CFGrote
JOSS Review: Documentation -- Example Usage
- [x] Provide a documented example where cgt is applied to real or realistic simulated data. Ideally, this would come as a jupyter notebook.
