### 2022-05-10 #4333 @CFGrote
3rd issue: https://github.com/jonathanHuwP/CrystalGrowthTracker/issues/22 regarding example use case
