### 2022-05-10 jonathanHuwP/CrystalGrowthTracker#24 @CFGrote
JOSS Review: Documentation -- Community guidelines
Community guidelines are missing. Please add (in README and doxygen build):
- [ ] Instructions how to contribute to the project
- [ ] Report issues or problems with the software
- [ ] Seek support
