### 2022-05-11 #4333 @CFGrote
General notes on the paper:
* I find parts of  it a bit hard to read and would recommend to have it cross-read by a native language speaker. I indicated some sentences that could be broken down into multiple sentences in attached annotated pdf.
* The paper is a bit sparse in citations. An interested reader may want to see one or two original papers where x-ray shadowgraphy has been applied to measure differential crystal growth.
[10.21105.joss.04333_annotated_CFGrote_20220511.pdf](https://github.com/openjournals/joss-reviews/files/8670191/10.21105.joss.04333_annotated_CFGrote_20220511.pdf)
