### 2022-05-30 #4333 @lucydot
Hello @jonathanHuwP, please see the comments above from @CFGrote. They have raised a number of issues on the software repo ([#20](https://github.com/jonathanHuwP/CrystalGrowthTracker/issues/20), [#21](https://github.com/jonathanHuwP/CrystalGrowthTracker/issues/21), [#22](https://github.com/jonathanHuwP/CrystalGrowthTracker/issues/22), [#24](https://github.com/jonathanHuwP/CrystalGrowthTracker/issues/24)) and have made some general comments [here](https://github.com/openjournals/joss-reviews/issues/4333#issuecomment-1123743505).

The review process is an iterative procedure (a conversation back and forth between reviewer and author); the review and any responses to the review run concurrently. So now is the time to start working on responses/improvements to the code after the comments from @CFGrote. Please update us on any progress / delays to progress in this thread.

For more information there is the [JOSS documentation](https://joss.readthedocs.io/en/latest/submitting.html#the-review-process), please also continue to ask me questions here if needed.
