### 2022-05-30 9691783
first 7 corrections on paper
