+++ paper.bib
@@ -18,3 +18,13 @@ doi = {10.1016/j.crhy.2012.10.010}
   volume = {31},
   number = {4},
 }
+
+@article{mvc88krasne,
+author = {Krasner, Glenn and Pope, Stephen},
+year = {1988},
+month = {01},
+pages = {},
+title = {A Description of the Model-View-Controller User Interface Paradigm in the Smalltalk80 System},
+volume = {1},
+journal = {Journal of Object-oriented Programming - JOOP}
+}
