### 2022-06-01 jonathanHuwP/CrystalGrowthTracker#20 @jonathanHuwP
The cmd tool inputs are now appearing as code and the other typos are corrected. The run_tests.py script runs all the unit tests and can direct the output to a CSV file, if required.
