### 2022-06-08 #4333 @lucydot
@JoannaLeng, thanks for your update.

Once you and @jonathanHuwP have a rough timescale for working through the reviewer points raised, please let us know here. The JOSS process is different from more traditional journals - it is a conversation between authors, reviewers and editors. So if anything is unclear or would benefit from more discussion please don't hesitate to raise it here in the issue thread. You don't need to address all the points in one go and then summarise here, we can discuss as working through.
