+++ paper.bib
@@ -17,6 +17,7 @@ doi = {10.1016/j.crhy.2012.10.010}
   year = {2012},
   volume = {31},
   number = {4},
+  doi={10.1145/2185520.2185561}
 }
 
 @article{mvc88krasne,
