### 2022-06-09 #4333 @JoannaLeng
@lucydot
@jonathanHuwP and I are just finishing our meeting. We will give some updates to you and the reviewers.
There is an issue with the references we wanted to ask your advice on. We were asked to provide a reference for the Model-View-Controller paridigm. The reference we found is from 1988 for which there is no DOI. This is likley to come up with an error. Is there a way we can suppess the error.
Also we reference a computer company and software for which there is no academic reference. Sometimes I use a reference to the homepage and reference that way. In this paper we have put hyper-links in the paper. Does JOSS have a prfered way to handle this type of reference?
