### 2022-06-09 #4333 @JoannaLeng
@KedoKudo and @CFGrote thank you for your reviews.

Both of you highlighted issues arround documentation and tutorials. @CFGrote created issues #22 and #23 that both relate to these concerns. We have looked into them and I have put nots on both of these issues below. Please let us know if you are happy with this solution.

JOSS Review: Documentation -- Example Usage #22
We intend to produce videos that will be placed into our account on YouTube for the documentation. We have chosen this over written instructions with screen shots or a jupyter notebook. We have picked this because 1,) the user must work interactively with the Crystal Growth Tracker and this changes with time and is easier to record with video rather than still screen shots 2,) We think that the Windowing system that comes with the GUI of the Crystal Growth Tracker will not work well with jupyter.

We are in the process of placing a test video of Shadowgraphs on Zenodo. We did not create the data and are getting meta data from the scientist who did create it so that we can properly credit them.

It would be best to do this now once we have resolved the reviewers concerns so the videos are up to date.

JOSS Review: Documentation -- Functionality documentation #23

This is a GUI application only.

Scientific videos are different to videos of a Hollywood film and most python video processing libraries are designed for Hollywood style videos. The scientists are interested in the time changing elements of the video and these are very subtle so that it is not possible to automate the detection of the important features and that means there is not enough data yet for AI deep learning techniques. The scientists only got less than a hundred useful frames of video from a few hundred of videos each with a few thousand frames!

We are looking into developing tutorials as described in issue #22
