### 2022-06-09 #4333 @JoannaLeng
@CFGrote

JOSS REVIEW: README.md #20
There were a number of issues here. Most were easy to fix except the unitest one. We closed that issue and created issue #25 just for the unitests.

For the other issues can you let us know how we decided when JOSS review issue are resolved.
