### 2022-06-09 #4333 @lucydot
> For the other issues can you let us know how we decided when JOSS review issue are resolved.

@JoannaLeng I think this may be a question about the review process, although I don't fully understand the wording. Are you asking how we determine if each of the review criteria are met? Each reviewer will work through the individual checklists they have generated in the thread above. Once they are persuaded a checklist item has been met (either through inspection of the repo, or through conversation with yourself, or by reading the paper) they will tick it off. Of course there will sometimes be differences in interpretation, in which case I can help to move the process along.  You may find the [reviewer guidelines](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html) for JOSS useful.
