### 2022-06-13 #4333 @jonathanHuwP
Apologies for not responding sooner. I didn't realize that this issue was the preferred means of communicationa and was wondering why no emails.

I have renamed the default branch from 'master' => 'main', you will need to either reclone or in you local repos, make sure you are using the old 'master' branch and run the following.

git branch -m master main
git fetch origin
git branch -u origin/main main
git remote set-head origin -a
