### 2022-06-13 e108796
removed debugging output
