### 2022-06-14 jonathanHuwP/CrystalGrowthTracker#21 @jonathanHuwP
I have added a setup.py file that allows pip install -e to run under windows.  I will go on and test using Ubuntu.
