### 2022-06-15 3003bab
moved test dir into cgt, and added encodings to open
