### 2022-06-16 f759b43
added main and flags for running tests
