### 2022-06-16 6065213
running main window and test for ffmpeg
