+++ cgt/main.py
@@ -1,5 +1,5 @@
 """
-Created on Sun 16 June 2022
+Created on Thur 16 June 2022
 
 Licensed under the Apache License, Version 2.0 (the "License"); you may not use
 this file except in compliance with the License. You may obtain a copy of the
