### 2022-06-16 3daf1a4
testing option now working in application main
