### 2022-06-16 d416859
removed debugging output & turned off stats log file
