### 2022-06-16 f095a0a
updated run_cgt.py to mirror main
