+++ cgt/gui/markupwidget.py
@@ -256,7 +256,7 @@ class MarkUpWidget(qw.QWidget, Ui_MarkUpWidget):
         self._cloneControls.set_frame_currently_displayed(self._current_frame)
 
         delay = self._video_source.get_video_data().get_user_time_step()
-        qc.QTimer.singleShot(delay, self.next_pixmap)
+        qc.QTimer.singleShot(int(delay*1000), self.next_pixmap)
 
     @qc.pyqtSlot()
     def next_pixmap(self):
