+++ environment_CGT.yml
@@ -4,7 +4,7 @@ channels:
   - conda-forge
   - anaconda
 dependencies:
-  - python=3.8
+  - python>3.8
   - pyqt
   - doxygen
   - pylint
