### 2022-06-17 41c14af
Create contributing.md
Added to ease community work.
