+++ contributing.md
@@ -1,0 +1,29 @@
+# How to contribute to CrystalGrowthTracker
+
+We want to thank you for your interest in contributing! The following guidelines to help you better understand how to best contribute to the project.
+
+## Coding conventions
+
+The main convention we observe it that all code should have a pylint score above 7.5.
+
+## You have found a bug?
+
+Check if the bug has already been reported in the project Issues, on GitHub.
+
+If you can't find an open issue, please open a new one.
+
+Be sure to include a meaningful title, followed by a clear description of the problem. Please give enough details for the developers to be able to recreate the problem. Also includ details of hardware, OS, and the python version (`python --version`) used.
+
+## If you have written a patch to fix the bug?
+
+Thanks for your efforts! Please submit a new GitHub pull request with the patch, and please only one patch in a pull request.
+
+Please ensure the description in the pull request clearly describes the problem and solution, see the above section.
+
+Before submitting new code, please note the coding convention, which is a pylint score above 7.5.
+
+## Do you want to add, or change, a feature?
+
+We would like to hear your ideas!
+
+Open an issue to describe your new feature, or how an existing feature should be changed, which will enable you to discuss the issue with other developers.
