### 2022-06-17 jonathanHuwP/CrystalGrowthTracker#21 @jonathanHuwP
I have added a main.py to cgt and specified it as an entry point in setup.py. Now when the package is installed with pip, it can run from the command line as cgt, with the same command line options as the run_cgt.py script. An extra option --test has been added that runs all the unittests and the ffmpeg tests.

I have successfully tested the package under Ubuntu 20 updated the README file.

The old run_unittests.py file has been renameded run_unittests_2_csv.py, which is more descriptive as it can save failed tests to a comma seperated value file.
