+++ paper.bib
@@ -33,7 +33,7 @@ journal = {Journal of Object-oriented Programming - JOOP}
  @misc{web:ffmpeg,
  title={FFmpeg},
  howpublished={Web site for opensource project},
- note = {Web site last visited 20 June 2022.}
+ note = {Web site last visited 20 June 2022.},
  url={https://www.ffmpeg.org/},
  author={Fabrice Bellard, originator of the FFmpeg project.}
  }
@@ -41,7 +41,7 @@ journal = {Journal of Object-oriented Programming - JOOP}
  @misc{web:riverbank,
  title={Riverbank Computing Limited},
  howpublished={Corperate web site.},
- note = {Web site last visited 20 June 2022.}
+ note = {Web site last visited 20 June 2022.},
  url={https://riverbankcomputing.com/},
  author={Riverbank Computing Limited}
  }
