+++ paper.bib
@@ -1,12 +1,12 @@
 @article{Baruchel2013,
-author = {Baruchel, José and Di Michiel, M. and Lafford, Tamzin and Lhuissier, Pierre and Meyssonnier, Jacques and Nguyen-Thi, Henri and Philip, Armelle and Pernot, Petra and Salvo, Luc and Scheel, Mario},
-year = {2013},
-month = {02},
-pages = {208–220},
-title = {Synchrotron X-ray imaging for crystal growth studies},
-volume = {14},
-journal = {Comptes Rendus Physique},
-doi = {10.1016/j.crhy.2012.10.010}
+  author = {Baruchel, José and Di Michiel, M. and Lafford, Tamzin and Lhuissier, Pierre and Meyssonnier, Jacques and Nguyen-Thi, Henri and Philip, Armelle and Pernot, Petra and Salvo, Luc and Scheel, Mario},
+  year = {2013},
+  month = {02},
+  pages = {208–220},
+  title = {Synchrotron X-ray imaging for crystal growth studies},
+  volume = {14},
+  journal = {Comptes Rendus Physique},
+  doi = {10.1016/j.crhy.2012.10.010}
 }
 
 @article{Wu12Eulerian,
@@ -21,27 +21,26 @@ doi = {10.1016/j.crhy.2012.10.010}
 }
 
 @article{mvc88krasne,
-author = {Krasner, Glenn and Pope, Stephen},
-year = {1988},
-month = {01},
-pages = {},
-title = {A Description of the Model-View-Controller User Interface Paradigm in the Smalltalk80 System},
-volume = {1},
-journal = {Journal of Object-oriented Programming - JOOP}
+  author = {Krasner, Glenn and Pope, Stephen},
+  year = {1988},
+  month = {01},
+  title = {A Description of the Model-View-Controller User Interface Paradigm in the Smalltalk80 System},
+  volume = {1},
+  journal = {Journal of Object-oriented Programming - JOOP}
 }
 
- @misc{web:ffmpeg,
- title={FFmpeg},
- howpublished={Web site for opensource project},
- note = {Web site last visited 20 June 2022.},
- url={https://www.ffmpeg.org/},
- author={Fabrice Bellard, originator of the FFmpeg project.}
- }
+@misc{web:ffmpeg,
+  title={FFmpeg},
+  howpublished={Opensource project web site.},
+  note = {Web site last visited 20 June 2022.},
+  url={https://www.ffmpeg.org/},
+  author={Fabrice Bellard (project originator)}
+}
 
- @misc{web:riverbank,
- title={Riverbank Computing Limited},
- howpublished={Corperate web site.},
- note = {Web site last visited 20 June 2022.},
- url={https://riverbankcomputing.com/},
- author={Riverbank Computing Limited}
- }
+@misc{web:riverbank,
+  title={Riverbank Computing Limited},
+  howpublished={Corperate web site.},
+  note = {Web site last visited 20 June 2022.},
+  url={https://riverbankcomputing.com/},
+  author={Riverbank Computing Limited}
+}
