### 2022-06-21 69e4443
moved test script to tests, removed old run script, linted
