+++ auto_pylint.py
@@ -242,7 +242,6 @@ def check_file(file_name):
     if path.exists() and path.is_file():
         handle = None
         try:
-            print("try handle")
             handle = open(path, 'w', encoding="UTF-8")
         except PermissionError:
             return None
@@ -253,7 +252,6 @@ def check_file(file_name):
 
     # create file with write permissions
     try:
-        print("try touch")
         path.touch(stat.S_IWUSR)
     except PermissionError:
         return None
