### 2022-06-21 1526841
udated Doxyfile and more corrections to README
