+++ paper.md
@@ -14,7 +14,7 @@ authors:
     orcid: 0000-0001-5283-8065
   - name: Sven L. M. Schroeder
     affiliation: "4,5,6,7"
-  - name: Gunjen Das
+  - name: Gunjan Das
     affiliation: "4,5,6,7"
 affiliations:
  - name: EPSRC Research Fellow in Software Engineering
@@ -41,7 +41,7 @@ X-ray synchrotron radiation allows the investigation of many physical processes 
 
 # Statement of Need
 
-Much of the output of global fine chemicals industries consists of crystalline powders produced by precipitation from solution.  Since the geometry of a crystal is the result of differential growth rates on different crystal planes, there is great commercial interest in studying the growth rates of crystal faces in the early stages of crystallization.  One approach is to use the X-rays of synchrotron radiation [@Baruchel2013] to produce shadowgraphs of crystals precipitating onto a substrate.  Videos of the shadow graphs first need enhancement using the packages such as Euler’s Magnifier [@Wu12Eulerian] and some statistical analysis of the raw video is desirable to find regions of interest.  **CrystalGrowthTracker** has been developed to assist in the analysis of videos of the raw and enhanced videos.
+Much of the output of global fine chemicals industries consists of crystalline powders produced by precipitation from solution.  Since the geometry of a crystal is the result of differential growth rates on different crystal planes, there is great commercial interest in studying the growth rates of crystal faces in the early stages of crystallization.  One approach is to use the X-rays of synchrotron radiation [@Baruchel2013] to produce shadowgraphs of crystals precipitating onto a substrate.  Videos of the shadow graphs first need enhancement using the packages such as Euler’s Magnifier [@Wu12Eulerian] and some statistical analysis of the raw video is desirable to find regions of interest.  **CrystalGrowthTracker** has been developed to assist in the analysis of videos of the raw and enhanced videos.  The package was developed by Leng and Pickering, with specifications provided by Schroeder, and tested by Das, who prsented on the project [@das:bca21,@das:isic21].
 
 # Design
 
