+++ paper.bib
@@ -44,3 +44,20 @@
   url={https://riverbankcomputing.com/},
   author={Riverbank Computing Limited}
 }
+
+@misc{das:bca21,
+  author={Gunjan Das},
+  organization={BCA-BACG Joint Spring Meeting 2021},
+  url={https://www.crystallography.org.uk/spring-meetings/},
+  howpublished={Online event},
+  year={2021},
+  title={Time-Resolved X-ray Phase Contrast Imaging of Continuous Anti-Solvent Crystallisation}
+}
+
+@misc{das:isic21,
+  author={Gunjan Das},
+  organization={ISIC 21 - 21st International Symposium on Industrial Crystallization},
+  year={2021},
+  howpublished={Virtual Event},
+  title={Time-Resolved X-ray Phase-Contrast Imaging of Continuous Anti-Solvent Crystallisation}
+}
