+++ paper.bib
@@ -31,16 +31,14 @@
 
 @misc{web:ffmpeg,
   title={FFmpeg},
-  howpublished={Opensource project web site.},
-  note = {Web site last visited 20 June 2022.},
+  howpublished={Opensource project web site, last visited 20 June 2022.},
   url={https://www.ffmpeg.org/},
-  author={Fabrice Bellard (project originator)}
+  author={Fabrice Bellard}
 }
 
 @misc{web:riverbank,
   title={Riverbank Computing Limited},
-  howpublished={Corperate web site.},
-  note = {Web site last visited 20 June 2022.},
+  howpublished={Corperate web site, last visited 20 June 2022.},
   url={https://riverbankcomputing.com/},
   author={Riverbank Computing Limited}
 }
