+++ paper.bib
@@ -59,3 +59,18 @@
   howpublished={Virtual Event},
   title={Time-Resolved X-ray Phase-Contrast Imaging of Continuous Anti-Solvent Crystallisation}
 }
+
+@inproceedings{10.1117/12.2530698,
+author = {Shashidhara Marathe and Malte Storm and Venkata S. C. Kuppilli and Richard Harrison and Gunjan Das and Sven L. M. Schroeder and Silvia Cipiccia and Florian Döring and Christian David and Christoph Rau},
+title = {{Development of synchrotron pink beam x-ray grating interferometer at the Diamond Light source I13-2 beamline}},
+volume = {11113},
+booktitle = {Developments in X-Ray Tomography XII},
+editor = {Bert Müller and Ge Wang},
+organization = {International Society for Optics and Photonics},
+publisher = {SPIE},
+pages = {269 -- 275},
+keywords = {x-ray grating interferometer, phase-contrast imaging, phase-stepping technique, Spatial Harmonic Imaging, X-ray source},
+year = {2019},
+doi = {10.1117/12.2530698},
+URL = {https://doi.org/10.1117/12.2530698}
+}
