### 2022-06-22 #4333 @lucydot
@jonathanHuwP thanks for the updates. To bring to your attention that there are some other outstanding points listed [here](https://github.com/openjournals/joss-reviews/issues/4333#issuecomment-1123743505) (regarding the paper) and [here](https://github.com/openjournals/joss-reviews/issues/4333#issuecomment-1118805990) (documentation and testing).

@CFGrote @KedoKudo pinging you here so that the progress @jonathanHuwP has made (see messages above) is brought to you attention.
