### 2022-06-23 jonathanHuwP/CrystalGrowthTracker#24 @jonathanHuwP
@CFGrote @JoannaLeng

I have added a contributing.md file giving instructions on how to:

1. contribute to the project
2. report bugs
3. suggest improvements
4. seek assistance.
