### 2022-06-23 jonathanHuwP/CrystalGrowthTracker#21 @jonathanHuwP
@CFGrote @JoannaLeng

1. The package runs from a main.py file (in directory cgt), which also provides a testing mode.
2. A setup.py is provided to allow installation via PIP.
3. The package has been tested on Ubuntu 20
4. The test_video.avi file is removed by the `finally` clause at line 47 of videosource_ffmpeg_test.py.
