### 2022-06-29 094d833
amend sys.path to allow script to run without PYTHONPATH
