### 2022-06-29 76c2df3
enabled props widget when props tab selcted
