### 2022-06-30 #4333 @lucydot
Hello, just to let you all know that I am taking some annual leave next week, so will be out of office until around the 7th July.
