+++ cgt/model/videoanalysisresultsstore.py
@@ -149,6 +149,7 @@ class VideoAnalysisResultsStore(qc.QObject):
             Throws:
                 IndexError: pop index out of range
         """
+        # TODO remove region properly
         markers = []
         markers.append(self.get_lines_for_region(index))
         markers.append(self.get_points_for_region(index))
