### 2022-07-06 c93466c
speed calculator moved to model
