### 2022-07-06 ee8cbec
printing displacement/time graphs
