+++ paper.bib
@@ -60,6 +60,22 @@
   title={Time-Resolved X-ray Phase-Contrast Imaging of Continuous Anti-Solvent Crystallisation}
 }
 
+@misc{zenodoSample,
+  title = {Time-Resolved X-ray Phase-Contrast Video Imaging of Continuous Anti-Solvent Crystallisation},
+  url = {https://doi.org/10.5281/zenodo.6801296},
+  author = {G Das},
+  howpublished={Online Data},
+  year=2022
+}
+
+@misc{installVid,
+  title = {Using the Crystal Growth Tracker (cgt)},
+  url = {https://youtu.be/wYYFnPkVBrY},
+  author = {J Leng},
+  howpublished = {YouTube},
+  year = 2022
+}
+
 @inproceedings{10.1117/12.2530698,
 author = {Shashidhara Marathe and Malte Storm and Venkata S. C. Kuppilli and Richard Harrison and Gunjan Das and Sven L. M. Schroeder and Silvia Cipiccia and Florian Döring and Christian David and Christoph Rau},
 title = {{Development of synchrotron pink beam x-ray grating interferometer at the Diamond Light source I13-2 beamline}},
