+++ paper.md
@@ -45,34 +45,35 @@ Much of the output of global fine chemicals industries consists of crystalline p
 
 # Design
 
-Although the obvious approach would be to use image analysis and machine learning, the relatively noisy data, the limited number of data sets and the need for verifiable results lead to a manual approach.  User analysis of video required the package to be based on a graphical user interface (GUI).  Finally the users would need to be able to download and run the package on any machine, so the package was developed using Python and PyQt5, from Riverbank Computing Limited [@web:riverbank], within the conda package management system.  The use of Qt naturally leads to an object orientated architecture.
+Although the obvious approach would be to use image analysis and machine learning, the relatively noisy data, the limited number of data sets and the need for verifiable results lead to a manual approach.  User analysis of video required the package to have a graphical user interface (GUI).  Finally the users needed to be able to download and run the package on any machine, so the package was developed using Python and PyQt5, from Riverbank Computing Limited [@web:riverbank], within the conda package management system.  The use of Qt naturally leads to an object orientated architecture.
 
-The package must allow the user to load two videos, the raw and the enhanced, and must provide the following functions.
+The package allows the user to load two videos, raw and enhanced, and provides the following functionality:
 
-1.  Analyse and display intensity statistics of the raw video, to assist selection of times of interest.
-2.  Play enhanced video and allow the selection of regions of interest.
-3.  In regions of interest allow features to be marked up on a given frame.  The markers being, lines for crystal faces, or points for other features.
-4.  Change the frame of a region of interest and move the markers to follow the underlying feature.
-5.  Use time and space calibration data to calculate the true speed of the marker motions.
-6.  Store the above data in an open human readable format and present a report in HTML.
+1.  Easy install and uninstall plus availablity of a translation package.
+2.  Analyse and display intensity statistics of the raw video, to assist selection of times of interest.
+3.  Play enhanced video and allow the selection of regions of interest.
+4.  In regions of interest allow features to be marked up on a given frame.  The markers being, lines for crystal faces, or points for other features.
+5.  Change the frame of a region of interest and move the markers to follow the underlying feature.
+6.  Use time and space calibration data to calculate the true speed of the marker motions.
+7.  Store the above data in an open human readable format and present a report in HTML.
 
-# Functionalities
+# Functionality
 
-A model-view-controller (MVC) software architecture [@mvc88krasne] was used, in which the view and the controller were merged into the main widget.  The video was accessed via a VideoSource object, which was held by the main window, together with project data, and results objects. These objects constitute the MVC architecture model and were made available to the widgets carrying out user functions by getter methods.  These widgets were themselves held by the main window in a tab widget.
+Videos documenting the [install](@installVid), [uninstall](@uninstallVid) and [operation](@operateVid) of the package are available on YouTube.  Sample video data is available via [Zenodo](@zenodoSample).
 
-The VideoSource object accessed the video via the ffmpeg package [@web:ffmpeg], which was encapsulated in a using the subprocess module. The frame access had to be in terms of time rather than frame number as when accessing a specified frame number ffmpeg scans the entire video to count frames from the start.  The widgets requiring video stored their own time in the video and when playing called for the frame at the next time step.  No effort was made to play at the correct number of frames per second.
+A model-view-controller (MVC) software architecture [@mvc88krasne] was used, in which the view and the controller were merged into the main widget.  The video was accessed via a VideoSource object, which was held by the main window, together with project data, and results objects. These objects constitute a MVC architecture and were made available to the widgets carrying out user functions by getter methods.  These widgets were themselves held by the main window in a tab widget.
 
[…]
