+++ paper.bib
@@ -70,6 +70,22 @@
 
 @misc{installVid,
   title = {Using the Crystal Growth Tracker (cgt)},
+  url = {https://youtu.be/tjmPqGec1vs},
+  author = {J Leng},
+  howpublished = {YouTube},
+  year = 2022
+}
+
+@misc{uninstallVid,
+  title = {Using the Crystal Growth Tracker (cgt)},
+  url = {https://youtu.be/aYjkYWifw4Q},
+  author = {J Leng},
+  howpublished = {YouTube},
+  year = 2022
+}
+
+@misc{operateVid,
+  title = {Using the Crystal Growth Tracker (cgt)},
   url = {https://youtu.be/wYYFnPkVBrY},
   author = {J Leng},
   howpublished = {YouTube},
