### 2022-07-12 #4333 @jonathanHuwP
@lucydot @CFGrote @KedoKudo
We think we have fixed all the issues raised.

1. The paper has been corrected, and the number of referances increaced.
2. The package can now be pip installed.
3. Community guidlines have been added.
4. Video documentation has been put on YouTube, showing examples of installation and usage.
5. Sample data has been put on Zenodo.
6. Acknowledgement have been added.
7. The roles of all the authors have been clarified.
8. master has been renamed main
9. we have clarified that the technique is strictly two dimensional.

Please let us know if there is anything more we need to do.

Best wishes
@jonathanHuwP
@JoannaLeng
