### 2022-07-14 jonathanHuwP/CrystalGrowthTracker#22 @CFGrote
Great usage instruction video https://www.youtube.com/watch?v=wYYFnPkVBrY !
