### 2022-07-14 jonathanHuwP/CrystalGrowthTracker#22 @CFGrote
In the "Trace Artifacts", the GUI crashes again when adding lines (after the 6th or 7th line).
```
$> cgt                                                       [joss-submission]
QXcbConnection: XCB error: 3 (BadWindow), sequence: 613, resource id: 17012649, major code: 40 (TranslateCoords), minor code: 0
QXcbConnection: XCB error: 3 (BadWindow), sequence: 2641, resource id: 17022700, major code: 40 (TranslateCoords), minor code: 0
QXcbConnection: XCB error: 3 (BadWindow), sequence: 2860, resource id: 16889909, major code: 40 (TranslateCoords), minor code: 0
QXcbConnection: XCB error: 3 (BadWindow), sequence: 3418, resource id: 16902349, major code: 40 (TranslateCoords), minor code: 0
QXcbConnection: XCB error: 3 (BadWindow), sequence: 3541, resource id: 16916071, major code: 40 (TranslateCoords), minor code: 0
Traceback (most recent call last):
  File "/home/grotec/Repositories/JOSSReviews/CrystalGrowthTracker/cgt/gui/markupview.py", line 291, in mouseReleaseEvent
    self.finish_line(self.mapToScene(event.pos()))
  File "/home/grotec/Repositories/JOSSReviews/CrystalGrowthTracker/cgt/gui/markupview.py", line 339, in finish_line
    self.extend_line(point)
  File "/home/grotec/Repositories/JOSSReviews/CrystalGrowthTracker/cgt/gui/markupview.py", line 329, in extend_line
    line = self._draw_line.line()
AttributeError: 'NoneType' object has no attribute 'line'
zsh: abort      cgt
```
