### 2022-07-14 #4333 @CFGrote
Paper is fine IMHO, would be great if the GUI crashes i reported in https://github.com/jonathanHuwP/CrystalGrowthTracker/issues/22 can be addressed.
