### 2022-07-14 #4333 @CFGrote
> Thanks @CFGrote I can see all your review boxes are ticked, can you confirm that the issue raised in https://github.com/jonathanHuwP/CrystalGrowthTracker/issues/22 is the only remaining acceptance-blocker?

Confirmed

>
> @KedoKudo please could you update any of your review tick boxes in light of the recent changes made by the authors?
>
> Please note that I am going away on annual leave for one week starting tomorrow (15th July).
