### 2022-07-18 jonathanHuwP/CrystalGrowthTracker#22 @jonathanHuwP
@CFGrote @JoannaLeng @KedoKudo

1. I have removed the call to the deleted 'display_error' function. So, failure to probe the video will generate a warning message box rather than crashing the program.
2. Regarding the second bug: I have placed a guard on the offending line, to prevent execution if a line object has not been created.  I was only able to generate the error twice myself, both times running PIP installed on Linux.  Since adding the guard I have not been able to replicate the error.
3. All existing tests pass on linux and windows.

Best wishes
@jonathanHuwP
