### 2022-07-19 #4333 @jonathanHuwP
@CFGrote @KedoKudo @lucydot @JoannaLeng

Hello,

Just to let you know, I posted on issue #22 last night.

Also I will be on leave from Thursday 21 July to Friday 5 August inclusive. I'll be abel to restart work on Monday 8 August.

Regards
   Jonathan
