### 2022-07-25 #4333 @lucydot
@CFGrote it appears that the points raised in https://github.com/jonathanHuwP/CrystalGrowthTracker/issues/22 have been addressed - however you will be best placed to judge if this is the case.
