### 2022-08-23 jonathanHuwP/CrystalGrowthTracker#22 @CFGrote
Small related issue: The link to youtube videos and sample video in the paper.md (both gh rendered version and downloaded pdf) don't work.
