### 2022-08-23 jonathanHuwP/CrystalGrowthTracker#22 @CFGrote
The observed crash is not appearing anymore, so can close this issue.
