### 2022-08-31 #4333 @lucydot
Hello, I can see that @CFGrote is happy that https://github.com/jonathanHuwP/CrystalGrowthTracker/issues/22 has been addressed, and all boxes in each checklist are checked - so I think we are good to go ahead with publication @CFGrote  @KedoKudo?
