### 2022-08-31 #4333 @lucydot
Hello, some corrections/suggestions for the paper @JoannaLeng @jonathanHuwP :

- I could not see the following from the checklist in the paper: `State of the field: Do the authors describe how this software compares to other commonly-used packages?`
-  typo: `charaterizing`
- re-wording: `, and beta tested by Das, who prsented on the project (Das, 2021a),(Das,
30 2021b).` --> `The package was developed by Leng and Pickering, with specifications provided by Schroeder. Das has beta tested and presented on the project [citations].`
- clarification: what is meant by the phrase ` Details of the data collection can be found in (Marathe et al., 2019).` ? Is this the data collection required to develop and test the software?
- Citation entry for Riverbank computing is not correct. You most probably need quotation marks around the full company name.
- re-wording: `A test suite is provided, that utilizes the Python unittest module` --> `A test suite based on the Python unittest module is provided`
- the links `install`, `uninstall`,`operation` and `zenodo` are not working.
- Would you consider adding a screenshot from the GUI? Not always useful, but in some cases it gives a quick flavour / indicator as to what a user would expect.
- I'll also note that the primary audience for this paper is software users rather than software developers - there is more information re: the technologies underlying the software than usual for a JOSS submission. Something to keep in mind when doing final edits. I won't suggest any massive changes.
