### 2022-09-01 d21d894
fix Lucy bug #2
