+++ paper.bib
@@ -37,10 +37,10 @@
 }
 
 @misc{web:riverbank,
-  title={Riverbank Computing Limited},
+  title={"Riverbank Computing Limited"},
   howpublished={Corperate web site, last visited 20 June 2022.},
   url={https://riverbankcomputing.com/},
-  author={Riverbank Computing Limited}
+  author={"Riverbank Computing Limited"}
 }
 
 @misc{das:bca21,
