+++ paper.bib
@@ -40,7 +40,7 @@
   title={"Riverbank Computing Limited"},
   howpublished={Corperate web site, last visited 20 June 2022.},
   url={https://riverbankcomputing.com/},
-  author={"Riverbank Computing Limited"}
+  author={{Riverbank Computing Limited}}
 }
 
 @misc{das:bca21,
