### 2022-09-01 77f554a
fix all hyprelinks
