+++ paper.md
@@ -61,7 +61,7 @@ The package allows the user to load two videos, raw and enhanced, and provides t
 
 # Functionality
 
-Videos documenting the [install](@installVid), [uninstall](@uninstallVid) and [operation](@operateVid) of the package are available on YouTube.  Sample video data is available via \href{https://doi.org/10.5281/zenodo.6801296}{Zenodo}.
+Videos documenting the \href{https://youtu.be/tjmPqGec1vs}{install}, \href{https://youtu.be/aYjkYWifw4Q}{uninstall} and \href{https://youtu.be/wYYFnPkVBrY}{operation} of the package are available on YouTube.  Sample video data is available via \href{https://doi.org/10.5281/zenodo.6801296}{Zenodo}.
 
 A model-view-controller (MVC) software architecture [@mvc88krasne] was used, in which the view and the controller were merged into the main widget.  The video was accessed via a VideoSource object, which was held by the main window, together with project data, and results objects. These objects constitute a MVC architecture and were made available to the widgets carrying out user functions by getter methods.  These widgets were themselves held by the main window in a tab widget.
 
