### 2022-09-01 90bfc9a
added state of field and imaged dir
