### 2022-09-01 9caec5c
added caption to image in README
