+++ README.md
@@ -10,12 +10,10 @@ Licensed under the Apache License, Version 2.0 (the "License"); you may not use 
 
 Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the specific language governing permissions and limitations under the License.
 
+|:--:|
 |![Screenshot of project details after loading a video file.](images/CGT_properties_tab.png?raw=true)|
-|:--:|
 | <b>Screenshot of project details after loading a video file.</b>|
-|:--:|
 |![Screenshot after statistical analysis a video file.](images/CGT_stats_tab.png?raw=true)|
-|:--:|
 | <b>Screenshot after statistical analysis a video file.</b>|
 
 ## Developed With:
