+++ README.md
@@ -16,6 +16,8 @@ Unless required by applicable law or agreed to in writing, software distributed 
 | <b>Screenshot of project details after loading a video file.</b>|
 |![Screenshot after statistical analysis a video file.](images/CGT_stats_tab.png?raw=true)|
 | <b>Screenshot after statistical analysis a video file.</b>|
+|![Screenshot during region selection.](images/CGT_select_tab.png?raw=true)|
+| <b>Screenshot during region selection.</b>|
 
 ## Developed With:
 This was developed using Python 3.8/3.9/3.10 and Anaconda, on Windows 10 systems, and also successfully tested on Ubuntu 20. The software has not been tested under any other operating systems.
