### 2022-09-01 #4333 @jonathanHuwP
@lucydot @JoannaLeng

Thank you for your input, Joanna and myself have worked through these points. Most changes will be easy to find, but the first point we have added a paragraph to the "Statement of Need" section.

Please can you clarify the last point as we are interested in submitting another paper. Does JOSS generally publish papers primarily for users or developers?

If you are not happy with anything please let us know and we will try to make corrections.
