### 2022-09-05 #4333 @lucydot
@jonathanHuwP @JoannaLeng could you check if there are dois available for two of the references - Das, G. (2021b) and Krasner, G., & Pope, S. (1988)?

> Please can you clarify the last point as we are interested in submitting another paper. Does JOSS generally publish papers primarily for users or developers?

I'm not exactly sure how best to answer this question. JOSS is a developer-friendly journal, however the review process is much more focused on user experience (e.g. documentation, installation, functionality) rather than reviewing the experience as a potential developer (e.g. digging into the code and how it is structured) - so I'd say the primary audience is users.

It looks to me that the numbered list of functionality could better sit under the "Functionality" heading. I don't think most of the existing text under the heading Functionality, or the section on Testing, Documentation and Linting is needed. If you would like to keep (part of) this text, it may sit better under the Design section. Happy to be challenged on this if you think differently!
