+++ paper.md
@@ -77,9 +77,11 @@ The package allows the user to load two videos, raw and enhanced, and provides t
 
 Project data and results were saved in comma-separated value (CSV) files using the Python csv module. CSV is a text format, which can be viewed and edited in text editors or spreadsheets, such as Excel.  The results of a project can be saved in a HTML report, which included tables, graphs and anotated video frames.
 
-# Testing, Documentation and Linting
+# For Developers
 
-A test suite based on the Python unittest module is provided.  Because the Python unittest module cannot run the ffmpeg subprocess a separate test was developed for these interactions.  The Qt QTest object is used to simulate IO events, and the capture the resultant signals.  The tests can be executed by adding the '-t' flag to the main cgt command line. The Python script run_test executes all tests and saves the results to a CSV file.
+Some readers may wish to extend this package, for them we provide some further information on testing, source code style and documentation.
+
+A test suite is provided, because the Python unittest module cannot run the ffmpeg subprocess a separate test was developed for these interactions.  The Qt QTest object is used to simulate IO events, and the capture the resultant signals for tesing the GUI.  The tests can be executed by adding the '-t' flag to the main cgt command line. The Python script run_test executes all tests and saves the results to a CSV file.
 
 Doxygen was used to generate documentation from source code comments.  The code was developed using the Pylint static code analysis tool, for which a runner script was developed. The script runs Pylint on all files, with the output displayed in the shell tool window or saved to CSV file.
 
