+++ paper.md
@@ -51,13 +51,11 @@ Videos of the shadow graphs first need enhancement using packages such as Eulerâ
 
 # Design
 
-Although the obvious approach would be to use image analysis and machine learning, the relatively noisy data, the limited number of data sets and the need for verifiable results lead to a manual approach.  User analysis of video required the package to have a graphical user interface (GUI).  Finally the users needed to be able to download and run the package on any machine, so the package was developed using Python and PyQt5, from Riverbank Computing Limited [@web:riverbank], within the conda package management system.  The use of Qt naturally leads to an object orientated architecture.
+Although the obvious approach would be to use image analysis and machine learning: the relatively noisy data; the limited number of data sets; and the need for verifiable results lead to a manual approach.  User analysis of video required the package to have a graphical user interface (GUI). The best way to co-ordinate a GUI with the computational engine that runs in the background is through the model-view-controller (MVC) software architecture (Krasner & Pope, 1988).Finally, the users needed to be able to download and run the package on any machine, so the package was developed using Python within the conda virtual environment and in the pip package management system.
 
-A model-view-controller (MVC) software architecture [@mvc88krasne] was used, in which the view and the controller were merged into the main widget.  The video was accessed via a VideoSource object, which was held by the main window, together with project data, and results objects. These objects constitute a MVC architecture and were made available to the widgets carrying out user functions by getter methods.  These widgets were themselves held by the main window.
+The PyQt5 toolkit, from Riverbank Computing Limited [@web:riverbank], was chosen for the GUI component as it runs on any machine, is feature rich, and implements cross thread message passing using slots and signals. The latter is necessary for the complex functionality needed to play the video forwards, backwards and jump to any frame/time with and without markers and zoom.
 
-The VideoSource object accessed the video via the ffmpeg package [@web:ffmpeg], which was encapsulated in a subprocess. The frame access had to be in terms of time rather than frame number as when accessing a specified frame number ffmpeg scans the entire video to count frames from the start.  The widgets requiring video stored their current video-play-time and when playing called for the frame at the next time-step.  No effort was made to play at the correct number of frames per second.
-
-The video was displayed using a QGraphicsView with an associated QGraphicsScene, which together provide a scene-graph and view. Video was displayed as a QPixmap at depth zero in the scene, and the user crystal-face-markers are QGrapicsItems at a depth with a positive value, above the pixmap. A zoom feature is provided by scaling the view matrix in the QGraphicsView.
+However the Qt video player did not allow the low level access to the video stream required for this analysis, so the ffmpeg package [@web:ffmpeg] was used to access the video files in the computational engine.  A custom viewer was constructed with Qt which allowed the video to be viewed in high resolution, with zoom, so markers and annotations could be viewed across the video and display how features changed over time.  No effort was made to play at the correct number of frames per second, as preserving the highest image resolution was considered paramount.
 
 The user was able to add a marker to an image feature in a single key-frame of the video, then advance to a new key-frame and drag a clone of the marker to the feature's new locations.  The markers were graphics items augmented to hold their frame numbers and the identity of their parent.  Calculating the speeds of motion consisted of ordering the chains of cloned graphics items by frame number, finding the displacements in pixel coordinates, converting the pixel distances to real distances using the scaling factor.  The pixels are assumed to be square as individual x-ray sensors in the array were square. Frame number intervals are converted to times and speeds calculated.
 
