+++ paper.md
@@ -51,7 +51,7 @@ Videos of the shadow graphs first need enhancement using packages such as Eulerâ
 
 # Design
 
-Although the obvious approach would be to use image analysis and machine learning: the relatively noisy data; the limited number of data sets; and the need for verifiable results lead to a manual approach.  User analysis of video required the package to have a graphical user interface (GUI). The best way to co-ordinate a GUI with the computational engine that runs in the background is through the model-view-controller (MVC) software architecture (Krasner & Pope, 1988).Finally, the users needed to be able to download and run the package on any machine, so the package was developed using Python within the conda virtual environment and in the pip package management system.
+Although the obvious approach would be to use image analysis and machine learning: the relatively noisy data; the limited number of data sets; and the need for verifiable results lead to a manual approach.  User analysis of video required the package to have a graphical user interface (GUI). The best way to co-ordinate a GUI with the computational engine that runs in the background is through the model-view-controller (MVC) software architecture (Krasner & Pope, 1988).  Finally, the users needed to be able to download and run the package on any machine, so the package was developed using Python within the conda virtual environment and in the pip package management system.
 
 The PyQt5 toolkit, from Riverbank Computing Limited [@web:riverbank], was chosen for the GUI component as it runs on any machine, is feature rich, and implements cross thread message passing using slots and signals. The latter is necessary for the complex functionality needed to play the video forwards, backwards and jump to any frame/time with and without markers and zoom.
 
