### 2022-09-22 #4333 @jonathanHuwP
@lucydot @JoannaLeng

Hello Lucy,

This is a list of what we have done to address your concerns.

•	The reference “Krasner, G., & Pope, S. (1988)” is from 1988 which is before the common use of DOIs and we have not been able to find any. This reference relates to the MVC (Model View Control” architecture that we used and mention in the design section. It is an old and well recognised architecture.

•	In 2021 our co-author Gunjan Das attended 2 online events where she gave presentations on the work she had done as part of her PhD which in both cases included a small section on the Crystal Growth Tracker software. Unfortunately, there were still Covid-19 issues and these meetings moved from face to face to online shortly before the events and the materials for the meeting do not seem to be as well archived as the previous events. Again, we have not been able to find a DOI.

•	We have moved the list of functionalities to the section with the heading Functionality.

•	We have concerns over other changes as we feel much of this content has come out of the review process and at the request of the reviewers, however when we wrote this, we were under the impression that the paper was more for a developer audience than a user audience. With this in mind we have edited to keep the topics there but made them more suitable for users and with some basic information for potential developers.
