### 2022-09-22 9907c42
Update paper.md
Some changes to grammar, and changing from past to present tense in the Design, Functionality and For Developers sections.
