### 2022-09-22 #4333 @choochoo-bot
Hi @jonathanHuwP @JoannaLeng,

Thanks for your response above - I think we can now go ahead to the final stages of publication. First, could you take a look at the PR I have submitted - it has some changes to grammar - please check that your meaning is not lost with what I have proposed. ---> https://github.com/jonathanHuwP/CrystalGrowthTracker/pull/27

Thanks also to @KedoKudo and @CFGrote for your reviews ✨
