### 2022-09-26 #4333 @lucydot
OK @jonathanHuwP the pdf looks good to me - final steps now. Please can you do the following:

- Make a tagged release of your software
- Archive this release on Zenodo (or similar)
    - The archive should have exactly the same title and author list as your JOSS paper
- Report the archive DOI in this thread
