### 2022-10-19 #4333 @lucydot
@jonathanHuwP the archive should have exactly the same title and author list as your JOSS paper - so in this case, the archive should be titled "CrystalGrowthTracker: A Python package to analyse crystal face advancement rates from time lapse synchrotron radiography" instead of "jonathanHuwP/CrystalGrowthTracker: JOSSPublication". It also needs to have the exact same author list as the JOSS paper (and not include choochoo-bot in the author list) - any questions, please let me know.
