### 2022-10-25 #4333 @jonathanHuwP
@lucydot @JoannaLeng

We have made 4 changes to the paper fixing typos, they are:

- In the job affiliations change EPSRC to senior for Joanna Leng so “EPSRC Research Fellow in Software Engineering” becomes “senior Research Fellow in Software Engineering”.
- Change is to are so “image analysis packages such as ImageJ, popular with scientific users, is designed” to “image analysis packages such as ImageJ, popular with scientific users, are designed”
- Add the words “or PDF” to “The results of a project can be saved in a HTML report” to “The results of a project can be saved in a HTML or PDF report”
- Remove the so in the “For Developers” section “used to simulate IO events, and the capture” becomes “used to simulate IO events, and capture”

We are now ready to publish if you accept these changes.
