### 2022-10-31 #4333 @lucydot
@openjournals/bcm-eics please note that have been changes made to the manuscript after the `recommend-accept` label was added. The latest version has been compiled in the message just above.

I don't think we need to re-archive the repo as only small changes have been made to the manuscript, but please let me know if this is wrong @openjournals/bcm-eics .
