### 2022-11-02 #4333 @Kevin-Mattheus-Moerman
@jonathanHuwP  I have just read your paper and checked the archive on ZENODO and all seems in order. I have a few very minor points which you can consider below:

- [ ] You may remove the zip/postal-codes from your affiliations. You can also remove county/province names and just use the city name, e.g. the following is sufficient: `Institute name, city/town name, country`. Also we tend to recommend to spell out country names so to use United Kingdom instead.
- [ ] Recommendation: Perhaps consider rewriting `For example, Das originally took eighteen...` to `For example, one of the authors, Gunjan Das, originally took eighteen...`.
- [ ] Both CrystalGrowthTracker and **CrystalGrowthTracker** occur in the text. Perhaps highlight them the same. We recommend using the code style text `CrystalGrowthTracker`, implemented like this in markdown:
```markdown
`CrystalGrowthTracker`
```
- [ ] Perhaps a comma can be added after `...cannot run the ffmpeg subprocess`?
