+++ paper.md
@@ -21,15 +21,15 @@ affiliations:
    index: 1
  - name: Research Fellow in Software Engineering
    index: 2
- - name: School of Computer Science, University of Leeds, Leeds, LS2 9JT, UK
+ - name: School of Computer Science, University of Leeds, Leeds, United Kingdom
    index: 3
- - name: School of Chemical and Process Engineering, University of Leeds, LS2 9JT, UK
+ - name: School of Chemical and Process Engineering, University of Leeds, Leeds, United Kingdom
    index: 4
- - name: Diamond Light Source Ltd, Oxfordshire, OX11 0DE, UK
+ - name: Diamond Light Source Ltd, Oxfordshire, United Kingdom
    index: 5
- - name: Research Complex at Harwell, Oxfordshire, OX11 0FA, Uk
+ - name: Research Complex at Harwell, Oxfordshire, United Kingdom
    index: 6
- - name: EPSRC Centre for Innovative Manufacturing in Continuous Manufacturing and Advanced Crystallisation, University of Strathclyde, G1 1RD, UK
+ - name: EPSRC Centre for Innovative Manufacturing in Continuous Manufacturing and Advanced Crystallisation, University of Strathclyde, United Kingdom
    index: 7
 date: 15 October 2021
 bibliography: paper.bib
