### 2022-11-07 d0330bb
minor grammer and format changes
