### 2022-11-07 #4333 @jonathanHuwP
@Kevin-Mattheus-Moerman @JoannaLeng

Hello,

Sorry to take so long, I was ill.

The recomended changes have been made and are on the joss-submission and main branches, which are now synced.

Please let us know if there is anything else we need to do.
