### 2022-11-08 #4333 @lucydot
@jonathanHuwP @Kevin-Mattheus-Moerman the affiliations don't look quite correct to me as they currently include job titles, and a city name was missing. I've updated and created a PR request here: https://github.com/jonathanHuwP/CrystalGrowthTracker/pull/28 .

I can confirm the other [suggested changes](https://github.com/openjournals/joss-reviews/issues/4333#issuecomment-1299807121) have been made.

@jonathanHuwP please confirm the changes are correct and merge the PR if so.
