### 2022-11-11 #4333 @Kevin-Mattheus-Moerman
@jonathanHuwP when @lucydot said "PR request", she meant a GitHub pull request. She has basically created some proposed changes. By accepting the pull request you effectively merge the changes she made into your project. Here is the link again:
https://github.com/jonathanHuwP/CrystalGrowthTracker/pull/28 can you have a look? Let me know if you need more help.
