### 2022-11-11 #4333 @Kevin-Mattheus-Moerman
@jonathanHuwP perhaps merging the pull request is no longer needed if you made the same changes yourself too. A couple of points remain open, see above :point_up: , e.g. on the version tag for the ZENODO archive. Can you work on that?
Also, can you check if your affiliations are now as per our guidelines. Pardon my ignorance but is `Oxfordshire` a county/province designation or is town/city name? Please use only the town/city name.
